feat(ideas): 想法新建/编辑改用第三列主内容区，替代 modal
对齐技能/MCP 的直连详情模式：想法 section 激活时第三列由
idea-detail store 驱动，点列表行进入查看/编辑 pane，+ 进入
新建 pane，两个 dialog 删除。列表加前端分页（每页 20 条）。

- 新增 stores/idea-detail（detailTarget + mutationTick，换团队自动清空）
- 新增 panel/IdeaDetailPane（create/edit pane + 空态提示）
- IdeasView：行点击/新建走 store，选中行高亮，底部翻页条
- App.tsx：ideas section 接管主列（隐藏 TabBar/ChatPanel），与 teamShare 同构
- locales：删孤儿键 ideas.detail.description，新增 ideas.detailEmpty
